ci(docs): set up ci for releasing docs (#5)
## What this PR does / why we need it:
<!--
What code changes are made?
What problem does this PR addresses, or what feature this PR adds?
-->
This PR sets up a CI (Continuous Integration) process for releasing
documentation. It includes code changes to automate the process of
releasing docs, such as setting up actions, triggers, and resolving
issues with colors and images. This improves the efficiency and
consistency of the documentation release process.

<!-- REPLACE WITH CONTENT -->

## Which issue(s) this PR resolves:
<!--
Usage: `Resolves #<issue number>`, or `Resolves <link to the issue>`.
If PR is about `failing-tests`, please post the related tests in a
comment and do not use `Resolves`
-->
N/A

## Special notes for your reviewer:
<!-- Do you think reviewers should focus on any particular parts of
code? -->
N/A

## Additional documentation e.g., enhancement proposals, usage docs,
etc.:
<!-- This section can be blank if this pull request does not require a
release note. -->
N/A

---------

Signed-off-by: Mateusz Urbanek <mateusz.urbanek.98@gmail.com>